refactor: split notifications metrics service by ISP
