+++ backend/apps/notifications/src/infrastructure/metrics/index.ts
@@ -1,1 +1,4 @@
-export * from "./metrics.factory";
+export * from "./http-metrics.service";
+export * from "./job-metrics.service";
+export * from "./email-metrics.service";
+export * from "./registry.manager";
