+++ backend/apps/notifications/src/shared/ports/metrics.port.ts
@@ -1,4 +1,12 @@
-export interface IMetricsService {
+import { Registry } from "prom-client";
+
+export interface IRegistryManager {
+  getRegistry(): Registry;
+  getMetrics(): Promise<string>;
+  getContentType(): string;
+}
+
+export interface IHttpMetricsService {
   incrementHttpRequestCount(
     method: string,
     route: string,
@@ -14,10 +22,16 @@ export interface IMetricsService {
     route: string,
     statusCode: string
   ): () => number;
+}
+
+export interface IJobMetricsService {
   incrementJobEnqueuedCount(jobName: string): void;
   incrementJobProcessedCount(jobName: string): void;
   incrementJobFailedCount(jobName: string): void;
   recordJobProcessingDuration(jobName: string): () => number;
+}
+
+export interface IEmailMetricsService {
   incrementEmailDeliveryCount(providerName: string, emailType: string): void;
   incrementEmailDeliveryErrorCount(
     providerName: string,
@@ -27,6 +41,4 @@ export interface IMetricsService {
     providerName: string,
     emailType: string
   ): () => number;
-  getMetrics(): Promise<string>;
-  getContentType(): string;
 }
